refactor: extract MCP tool registry
